requirements.txt
langchain>=0.2.0
langchain-openai>=0.2.0
langchain_community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-chroma>=0.1.4
GitPython>=3.1.43
openai>=1.0.0
python-dotenv
pydantic>=2.0.0
requests
aiohttp
faiss-cpu        # ローカルに簡単にベクトル検索を実装
chromadb         # 軽量なローカルDB
